requirements.txt
beepy==1.0.7
dateutils==0.6.12
plyer==2.0.0
PyAudio==0.2.11
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
retrying==1.3.3
simpleaudio==1.0.4
six==1.15.0
ta==0.7.0
urllib3==1.26.4

